// apps/web/src/pages/dzent/lib/enums.ts
export enum SystemMessageActionsEnum {
  Actions = 'actions',
  Text = 'text',
  Number = 'number',
  Radio = 'radio',
  Checkbox = 'checkbox',
  SinglePicklist = 'single-picklist',
  MultiPicklist = 'multi-picklist',
  CustomRangeOptionsPicklist = 'custom-range-options-picklist',
  Date = 'date',
  Daterange = 'date-range',
  CustomQuestions = 'custom-questions',
  MultiFileUpload = 'multi-file-upload',
  SingleFileUpload = 'single-file-upload',
  CustomFields = 'custom-fields',
}

export enum SystemMessageInstructionsEnum {
  Text = 'text',
}

export enum UserMessageViewEnum {
  Text = 'text',
  Actions = 'actions',
  Radio = 'radio',
  Checkbox = 'checkbox',
  SinglePicklist = 'single-picklist',
  MultiPicklist = 'multi-picklist',
  CustomRangeOptionsPicklist = 'custom-range-options-picklist',
  Date = 'date',
  Daterange = 'date-range',
  CustomQuestions = 'custom-questions',
  Files = 'files',
  MultiFileUpload = 'multi-file-upload',
  SingleFileUpload = 'single-file-upload',
  CustomFields = 'custom-fields',
}